+++ lazybuildinfo/doc.go
@@ -1,0 +1,2 @@
+// Package lazybuildinfo exposes Go build metadata to development tooling.
+package lazybuildinfo